Pull Button extension into separate file
